extract url-moniker and sign-only crates
