+++ Cargo.toml
@@ -110,6 +110,7 @@ members = [
     "runtime",
     "runtime-transaction",
     "send-transaction-service",
+    "sign-only",
     "stake-accounts",
     "storage-bigtable",
     "storage-bigtable/build-proto",
@@ -139,6 +140,7 @@ members = [
     "unified-scheduler-logic",
     "unified-scheduler-pool",
     "upload-perf",
+    "url-moniker",
     "validator",
     "version",
     "vortexor",
@@ -494,6 +496,7 @@ solana-serialize-utils = "=2.2.1"
 solana-sha256-hasher = "=2.2.1"
 solana-signature = "=2.2.1"
 solana-signer = "=2.2.1"
+solana-sign-only = { path = "sign-only", version = "=2.2.0" }
 solana-slot-hashes = "=2.2.1"
 solana-slot-history = "=2.2.1"
 solana-time-utils = "=2.2.1"
@@ -544,6 +547,7 @@ solana-transaction-metrics-tracker = { path = "transaction-metrics-tracker", ver
 solana-turbine = { path = "turbine", version = "=2.2.0" }
 solana-type-overrides = { path = "type-overrides", version = "=2.2.0" }
 solana-udp-client = { path = "udp-client", version = "=2.2.0" }
+solana-url-moniker = { path = "url-moniker", version = "=2.2.0" }
 solana-validator-exit = "=2.2.1"
 solana-version = { path = "version", version = "=2.2.0" }
 solana-vote = { path = "vote", version = "=2.2.0" }
