+++ clap-utils/Cargo.toml
@@ -27,6 +27,8 @@ solana-remote-wallet = { workspace = true, features = ["default"] }
 solana-seed-phrase = { workspace = true }
 solana-signature = { workspace = true }
 solana-signer = { workspace = true }
+solana-sign-only = { workspace = true }
+solana-url-moniker = { workspace = true }
 thiserror = { workspace = true }
 tiny-bip39 = { workspace = true }
 uriparse = { workspace = true }
