+++ clap-v3-utils/Cargo.toml
@@ -29,6 +29,8 @@ solana-seed-derivable = { workspace = true }
 solana-seed-phrase = { workspace = true }
 solana-signature = { workspace = true }
 solana-signer = { workspace = true }
+solana-sign-only = { workspace = true }
+solana-url-moniker = { workspace = true }
 solana-zk-token-sdk = { workspace = true }
 thiserror = { workspace = true }
 tiny-bip39 = { workspace = true }
