+++ clap-v3-utils/src/input_validators.rs
@@ -1,3 +1,4 @@
+pub use solana_url_moniker::normalize_to_url_if_moniker;
 use {
     crate::{
         input_parsers::signer::{SignerSource, SignerSourceKind},
@@ -248,17 +249,6 @@ where
         }
         Err(err) => Err(format!("{err}")),
     }
-}
-
-pub fn normalize_to_url_if_moniker<T: AsRef<str>>(url_or_moniker: T) -> String {
-    match url_or_moniker.as_ref() {
-        "m" | "mainnet-beta" => "https://api.mainnet-beta.solana.com",
-        "t" | "testnet" => "https://api.testnet.solana.com",
-        "d" | "devnet" => "https://api.devnet.solana.com",
-        "l" | "localhost" => "http://localhost:8899",
-        url => url,
-    }
-    .to_string()
 }
 
 #[deprecated(
