+++ clap-v3-utils/src/keypair.rs
@@ -9,6 +9,7 @@
 //! sources supported by the Solana CLI. Many other functions here are
 //! variations on, or delegate to, `signer_from_path`.
 
+pub use solana_sign_only::SignOnly;
 use {
     crate::{
         input_parsers::signer::{try_pubkeys_sigs_of, SignerSource, SignerSourceKind},
@@ -43,23 +44,6 @@ use {
     },
 };
 
-pub struct SignOnly {
-    pub blockhash: Hash,
-    pub message: Option<String>,
-    pub present_signers: Vec<(Pubkey, Signature)>,
-    pub absent_signers: Vec<Pubkey>,
-    pub bad_signers: Vec<Pubkey>,
-}
-
-impl SignOnly {
-    pub fn has_all_signers(&self) -> bool {
-        self.absent_signers.is_empty() && self.bad_signers.is_empty()
-    }
-
-    pub fn presigner_of(&self, pubkey: &Pubkey) -> Option<Presigner> {
-        presigner_from_pubkey_sigs(pubkey, &self.present_signers)
-    }
-}
 pub type CliSigners = Vec<Box<dyn Signer>>;
 pub type SignerIndex = usize;
 pub struct CliSignerInfo {
