+++ cli-config/Cargo.toml
@@ -15,8 +15,8 @@ lazy_static = { workspace = true }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 serde_yaml = { workspace = true }
-solana-clap-utils = { workspace = true }
 solana-commitment-config = { workspace = true }
+solana-url-moniker = { workspace = true }
 url = { workspace = true }
 
 [dev-dependencies]
