+++ cli-output/src/cli_output.rs
@@ -21,7 +21,6 @@ use {
         encode_ui_account, parse_account_data::AccountAdditionalDataV3,
         parse_token::UiTokenAccount, UiAccountEncoding, UiDataSliceConfig,
     },
-    solana_clap_utils::keypair::SignOnly,
     solana_clock::{Epoch, Slot, UnixTimestamp},
     solana_epoch_info::EpochInfo,
     solana_hash::Hash,
@@ -32,6 +31,7 @@ use {
         RpcAccountBalance, RpcContactInfo, RpcInflationGovernor, RpcInflationRate, RpcKeyedAccount,
         RpcSupply, RpcVoteAccountInfo,
     },
+    solana_sign_only::SignOnly,
     solana_signature::Signature,
     solana_sysvar::stake_history::StakeHistoryEntry,
     solana_transaction::{versioned::VersionedTransaction, Transaction},
