+++ sign-only/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "solana-sign-only"
+description = "Solana SignOnly type. For CLI usage."
+documentation = "https://docs.rs/solana-sign-only"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+solana-hash = { workspace = true }
+solana-presigner = { workspace = true }
+solana-pubkey = { workspace = true }
+solana-signature = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
